requirements.txt
kafka-python==2.0.2

pymongo==4.3.3

pandas==1.5.2
numpy==1.23.5

dash==2.7.1
plotly==5.11.0

python-dateutil==2.8.2
pytz==2022.7 